requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
opencv-python-headless